refactor(widget): static method
static method
